Consolidate kiam configuration (#1354)
Consolidates the kiam parts into the kiamSupport configuration.
Also adds the listeners for server and agent as configurable options.
The previously KIAMImage type is now underneath kiamSupport, so this is a breaking change to existing cluster.yaml files

Now looks like this
```
  kiamSupport:
    enabled: true
    image:
      repo: quay.io/uswitch/kiam
      tag: v2.8
    server: localhost
    agent: kiam-server
```

The original DNS health check/server settings are preserved as defaults
The kiam image is bumped to v2.8

Resolves #1350

Conflicts:
	core/controlplane/config/config.go
	core/controlplane/config/templates/cluster.yaml
	test/integration/maincluster_test.go