feat: Ability to toggle node pool rolling strategy between Sequential and Parallel (#1428)
Adds ` nodePoolRollingStrategy: <string> (parallel/sequential)` under `worker` and `worker.nodePools[]`. The former is for setting a strategy for all pools. The latter allows finer-grained control, like parallelizing roll per group of node pools.

See the discussion in the PR introduced this commit #1428 to see more details and expected use-cases of this feature.